ReactGPT

Documentation for ReactGPT.

ReactGPT.ReactGPT

ReactGPT.AbstractAction

ReactGPT.MediaWikiAction

ReactGPT.WikipediaAction

ReactGPT.action_invocation

ReactGPT.describe_action

ReactGPT.execute_action

ReactGPT.ReactGPT — Module

ReactGPT

This package was inspired by https://til.simonwillison.net/llms/python-react-pattern, and is a simple implementation of the ReAct LLM pattern in Julia.

The idea is that we can give LLMs access to the Internet, or to a database, or really any other resource, by defining a set of actions that the LLM can take.

Theoretically, one could have an EvalAction which lets the LLM run arbitrary Julia code on your computer. Please don't do this - Julia isn't sandboxed, and the risk that the model accidentally runs rm("/"; force = true) or some similarly destructive action is really quite high.

Usage

Interface

We provide an AbstractAction interface which handles most of the stuff.

ReactGPT.AbstractAction — Type

AbstractAction

This is the abstract type for all actions that the LLM can use. Actions must satisfy the action interface, i.e., implement the following methods:

action_invocation(action::AbstractAction): returns a String describing how the action is invoked. In future, this should return a regex.

describe_action(action::AbstractAction): returns a string describing the action, which is fed into the intitial prompt to the LLM.

execute_action(action::AbstractAction, args::AbstractString): executes the action, and returns a String which is the response to that action.

ReactGPT.MediaWikiAction — Type

MediaWikiAction(; api_url = "https://en.wikipedia.org/w/api.php", name = "Wikipedia")

Query any MediaWiki website.

ReactGPT.WikipediaAction — Type

WikipediaAction()

Query Wikipedia directly.

ReactGPT.action_invocation — Method

action_invocation(action::AbstractAction)

Returns a String describing how the action is invoked. In future, this should return a regex.

ReactGPT.describe_action — Method

describe_action(action::AbstractAction)

Returns a string describing the action, which is fed into the intitial prompt to the LLM.

ReactGPT.execute_action — Method

execute_action(action::AbstractAction, args::AbstractString)

Executes the action, and returns a String which is the response to that action. The arguments must always be typed as AbstractStrings, in order to accomodate SubStrings parsed from the LLM's response.